[TEST] Move Static Base Cache tests from painter.test.ts to static base cache test file [QUAL] In style_layer, set unchangedFrameCount as a public attribute
