+++ src/render/painter.test.ts
@@ -100,86 +100,6 @@ describe('translucent cache', () => {
         painter.destroy();
     });
 
-    test('captureCache creates texture and copies framebuffer', () => {
-        const manager = painter.staticBaseCache;
-        const cache = manager._cache;
-        expect(cache._texture).toBeNull();
-        expect(cache._cachedLayerCount).toBe(0);
-
-        manager.captureCache(painter.context, painter.width, painter.height, 7);
-
-        expect(cache._texture).toBeTruthy();
-        expect(cache._width).toBe(512);
-        expect(cache._height).toBe(512);
-        expect(cache._cachedLayerCount).toBe(7);
-        expect(gl.copyTexSubImage2D).toHaveBeenCalledWith(
-            gl.TEXTURE_2D, 0, 0, 0, 0, 0, 512, 512
-        );
-    });
-
-    test('captureCache reuses texture when size unchanged', () => {
-        const manager = painter.staticBaseCache;
-        const cache = manager._cache;
-        manager.captureCache(painter.context, painter.width, painter.height, 3);
-        const firstTexture = cache._texture;
-
-        manager.captureCache(painter.context, painter.width, painter.height, 5);
-        expect(cache._texture).toBe(firstTexture);
-        expect(cache._cachedLayerCount).toBe(5);
-    });
-
-    test('captureCache recreates texture on resize', () => {
-        const manager = painter.staticBaseCache;
-        const cache = manager._cache;
-        manager.captureCache(painter.context, painter.width, painter.height, 3);
-        const firstTexture = cache._texture;
-        const destroySpy = vi.spyOn(firstTexture, 'destroy');
-
-        painter.width = 1024;
-        painter.height = 768;
-        manager.captureCache(painter.context, painter.width, painter.height, 4);
-
-        expect(destroySpy).toHaveBeenCalled();
-        expect(cache._texture).not.toBe(firstTexture);
-        expect(cache._width).toBe(1024);
-        expect(cache._height).toBe(768);
-    });
-
-    test('destroy cleans up cache texture', () => {
-        const manager = painter.staticBaseCache;
-        const cache = manager._cache;
-        manager.captureCache(painter.context, painter.width, painter.height, 3);
-        const destroySpy = vi.spyOn(cache._texture, 'destroy');
-
-        painter.destroy();
-        expect(destroySpy).toHaveBeenCalled();
-    });
-
-    test('_blitCacheToScreen binds texture and draws quad', () => {
-        const manager = painter.staticBaseCache;
-        const cache = manager._cache;
-        manager.captureCache(painter.context, painter.width, painter.height, 5);
-
-        const drawSpy = vi.fn();
-        vi.spyOn(painter, 'useProgram').mockReturnValue({draw: drawSpy} as any);
-
-        cache._blitCacheToScreen(painter);
-
-        expect(gl.bindTexture).toHaveBeenCalledWith(
-            gl.TEXTURE_2D, cache._texture.texture
-        );
-        expect(painter.useProgram).toHaveBeenCalledWith('fullscreenTexture', null, true);
-        expect(drawSpy).toHaveBeenCalledTimes(1);
-
-        const args = drawSpy.mock.calls[0];
-        expect(args[0]).toBe(painter.context);
-        expect(args[1]).toBe(gl.TRIANGLES);
-        expect(args[9]).toBe('$fullscreenTexture');
-        expect(args[10]).toBe(painter.viewportBuffer);
-        expect(args[11]).toBe(painter.quadTriangleIndexBuffer);
-        expect(args[12]).toBe(painter.viewportSegments);
-    });
-
     test('cache is valid when size matches and enough stable layers', () => {
         const manager = painter.staticBaseCache;
         const cache = manager._cache;
@@ -198,7 +118,7 @@ describe('translucent cache', () => {
                 id,
                 type: 'fill',
                 source: 'test',
-                _unchangedFrameCount: 10,
+                unchangedFrameCount: 10,
                 isHidden: () => false,
                 is3D: () => false,
                 hasOffscreenPass: () => false,
@@ -311,7 +231,7 @@ describe('translucent cache', () => {
                 id,
                 type: 'fill',
                 source: 'test',
-                _unchangedFrameCount: 10,
+                unchangedFrameCount: 10,
                 isHidden: () => false,
                 is3D: () => false,
                 hasOffscreenPass: () => false,
@@ -319,7 +239,7 @@ describe('translucent cache', () => {
                 hasActiveTransition: () => false,
             };
         }
-        mockLayers['layer-5']._unchangedFrameCount = 0;
+        mockLayers['layer-5'].unchangedFrameCount = 0;
 
         painter.style = {
             _order: layerIds,
