+++ src/style/style.ts
@@ -872,7 +872,7 @@ export class Style extends Evented {
                 layer.hasActiveTransition(currentTime) ||
                 (layer.source && (reloadedSources.has(layer.source) ||
                     this._sourcesWithTileUpdates.has(layer.source)));
-            layer._unchangedFrameCount = (layerChanged) ? 0 : layer._unchangedFrameCount + 1;
+            layer.unchangedFrameCount = (layerChanged) ? 0 : layer.unchangedFrameCount + 1;
         }
         this._sourcesWithTileUpdates.clear();
     }
