+++ src/style/style_layer.ts
@@ -113,7 +113,7 @@ export abstract class StyleLayer extends Evented {
 
     _visibilityExpression: VisibilityExpression;
 
-    _unchangedFrameCount: number = 0;
+    unchangedFrameCount: number = 0;
 
     readonly onAdd: ((map: Map) => void);
     readonly onRemove: ((map: Map) => void);
